edge/web/requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2
psycopg2-binary>=2.9
requests>=2.31
grpcio>=1.62
protobuf>=4.25
netifaces>=0.11
psutil>=5.9
qrcode[pil]>=7.4
whitenoise>=6.6
aiortc>=1.7,<2.0
opencv-python-headless==4.10.0.84